Add files via upload
Python does not have any libraries that facilitate performing principal component analysis (PCA). 
One of the utilities of PFA is making dimention reduction.
I gathered a bunch of funtions that takes a xlsx file, makes a pandas Dataframe, performs a PFA, selects the 3 vatiables that contribute the most and
prunes original data of all  variables but the 3 variables selected.